+++ CHANGELOG.md
@@ -10,3 +10,8 @@
 ## 1.0.7 (2024-11-23)
 - Added support for players names in match game
 - Added Direction.fromString
+
+## 1.0.9 (2024-11-24)
+- Added support for match name in match game
+- Added support for teams name in match game
+- fixes with players names in match game